+++ cpp/test/Ice/slicing/objects/AllTests.cpp
@@ -571,7 +571,7 @@ public:
         test(p3->pi == 3);
         for(int i = 0; i < 300; ++i)
         {
-            PCDerived2Ptr p2 = ICE_DYNAMIC_CAST(PCDerived2, p3->pbs[i]);
+            PCDerived2Ptr p2 = ICE_DYNAMIC_CAST(PCDerived2, p3->pbs[static_cast<size_t>(i)]);
             test(p2->pi == i);
             test(p2->pbs.size() == 1);
             test(!p2->pbs[0]);
@@ -2910,7 +2910,7 @@ allTests(Test::TestHelper* helper)
             test(p3->pi == 3);
             for(i = 0; i < 300; ++i)
             {
-                PCDerived2Ptr p2 = ICE_DYNAMIC_CAST(PCDerived2, p3->pbs[i]);
+                PCDerived2Ptr p2 = ICE_DYNAMIC_CAST(PCDerived2, p3->pbs[static_cast<size_t>(i)]);
                 test(p2->pi == i);
                 test(p2->pbs.size() == 1);
                 test(!p2->pbs[0]);
